move ObjectIdentifier to break an upcoming import cycle
